+++ resources/compiledJS/script/hotbar.js
@@ -8,6 +8,7 @@ class gameSettings {
     constructor(base) {
         this.log_enemy_movement = base.log_enemy_movement || false;
         this.toggle_minimap = base.toggle_minimap || true;
+        this.hide_helmet = base.hide_helmet || false;
         this.hotkey_inv = base.hotkey_inv || "i";
         this.hotkey_char = base.hotkey_char || "c";
         this.hotkey_perk = base.hotkey_perk || "p";
@@ -30,6 +31,7 @@ class gameSettings {
 let settings = new gameSettings({
     log_enemy_movement: false,
     toggle_minimap: true,
+    hide_helmet: false,
     hotkey_inv: "i",
     hotkey_char: "c",
     hotkey_perk: "p",
@@ -157,6 +159,7 @@ function generateHotbar() {
                 const abiDiv = document.createElement("div");
                 const abiImg = document.createElement("img");
                 abiDiv.classList.add("ability");
+                abiDiv.classList.add(abi.equippedSlot);
                 if (state.abiSelected == abi && state.isSelected) {
                     frame.style.border = "4px solid gold";
                 }
@@ -195,11 +198,35 @@ function generateHotbar() {
                         abiDiv.addEventListener("mouseleave", e => renderTileHover(player.cords, e));
                     }
                 }
+                dragElem(abiDiv, [hotbar], updateUI);
                 abiDiv.append(abiImg);
                 frame.append(abiDiv);
             }
         });
     }
+}
+function swapAbility(from, to) {
+    console.log(from);
+    console.log(to);
+    let mainSkill = null;
+    let replaceSkill = null;
+    player.inventory.map((itm) => {
+        if (itm.equippedSlot === from)
+            mainSkill = itm;
+        if (itm.equippedSlot === to)
+            replaceSkill = itm;
+    });
+    player.abilities.map((abi) => {
+        if (abi.equippedSlot === from)
+            mainSkill = abi;
+        if (abi.equippedSlot === to)
+            replaceSkill = abi;
+    });
+    if (mainSkill)
+        mainSkill.equippedSlot = to;
+    if (replaceSkill)
+        replaceSkill.equippedSlot = from;
+    updateUI();
 }
 function generateSummonList() {
     const summonList = document.querySelector(".playerUI .summonList");
@@ -228,6 +255,10 @@ function generateSummonList() {
     });
 }
 function useConsumable(itm) {
+    if (dragging) {
+        dragging = false;
+        return;
+    }
     if (itm.healValue) {
         player.stats.hp += itm.healValue;
         spawnFloatingText(player.cords, itm.healValue.toString(), "lime", 36, 1000, 200);