+++ Changelog.md
@@ -12,12 +12,15 @@
   - Ranger, a medium set for all characters.
 - Added `Hunter` artifact set.
 - Reintroduced stat randomness for artifacts.
+- Added the option to toggle helmet visibility.
 
 ## Fixes & Tweaks
 
 - Fixed duplicate items being removed when selling one copy.
 - Made some screens scale slightly better.
 - Reworked keyboard movement, added diagonal movement keys.
+- Fixed some player rendering errors.
+- Made hotbar items and abilities draggable.
 
 # 1.0.4 ~ 20.2.2022